+++ src/main/java/kr/hhplus/be/server/domain/point/repository/PointRepository.java
@@ -5,11 +5,11 @@ import kr.hhplus.be.server.domain.point.entity.Point;
 public interface PointRepository {
 
     //포인트 조회
-    Point getUserPoint(int userId);
+    Point getUserPoint(long userId);
 
     //잔액 충전
-    Integer chargePoint(int userId, int point);
+    Integer chargePoint(long userId, int point);
 
     //포인트 차감
-    Integer usePoint(int userId, int point);
+    Integer usePoint(long userId, int point);
 }
