+++ src/main/java/kr/hhplus/be/server/infra/coupon/jparepository/UserCouponJpaRepository.java
@@ -15,14 +15,14 @@ public interface UserCouponJpaRepository extends JpaRepository<UserCoupon, Long>
 
     // 사용자 쿠폰 목록 조회
     @Query("SELECT c FROM UserCoupon c WHERE c.userId = :userId")
-    List<UserCoupon> getUserCoupon(@Param("userId")int userId);
+    List<UserCoupon> getUserCoupon(@Param("userId")long userId);
 
     // 사용자 쿠폰 한행
     @Query("SELECT uc FROM UserCoupon uc WHERE uc.userId = :userId AND uc.couponId = :couponId")
-    UserCoupon getByUserId(@Param("userId")int userId, @Param("couponId") int couponId);
+    UserCoupon getByUserId(@Param("userId")long userId, @Param("couponId") long couponId);
 
     @Query("SELECT uc FROM UserCoupon uc WHERE uc.userId = :userId AND uc.couponId = :couponId")
-    Optional<UserCoupon> findByUserIdAndCouponId(@Param("userId")int userId, @Param("couponId") int couponId);
+    Optional<UserCoupon> findByUserIdAndCouponId(@Param("userId")long userId, @Param("couponId") long couponId);
 
     // 사용자 쿠폰 사용
     //@Lock(LockModeType.PESSIMISTIC_WRITE)
